Processing overhaul
- Generalized methods in extract_layers to make adding new pickers easier. New methods are harder to human read but have more repeatable behaviours.
- Added a new wavelet method that uses a gausian approach. Up to 10x faster than current TN method!
- Removed: laypicking.py was no longer needed to define picker arguments
- Merge: GSGC_NK.py and NSIDC.py were merged as GSFC.py
- Merge: Wavelet_JK and Wavelet_TN were merged as Wavelet.py
