+++ pySnowRadar/__init__.py
@@ -1,5 +1,4 @@
 name = 'pySnowRadar'
 
 from .snowradar import SnowRadar
-from .layerpicking import pick_layers
 from .atm import ATM
